Add feature to query and visualize data
Modified the Project. ipynb by adding queries and functions to get Uber trips' popularity each weekday, the 10 busiest days with an average distance of trips in those days in 2009, and the trips and weather data in the period of Hurricane Sandy.  Also, added data visualization including a bar plot of the average trip's distance in each month, the heatmap of pickup and dropoff locations, and a scatter of tips of taxi(both original and binned data) against hourly precipitation.